+++ src/main/java/uz/ccrew/matchmaking/controller/AuthController.java
@@ -6,7 +6,7 @@ import uz.ccrew.matchmaking.dto.Response;
 import uz.ccrew.matchmaking.dto.UserDTO;
 import uz.ccrew.matchmaking.dto.auth.LoginDTO;
 import uz.ccrew.matchmaking.dto.auth.LoginResponseDTO;
-import uz.ccrew.matchmaking.service.AuthService;
+import uz.ccrew.matchmaking.service.impl.AuthServiceImpl;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -24,23 +24,23 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @Tag(name = "Auth Controller", description = "Auth API")
 public class AuthController {
-    private final AuthService authService;
+    private final AuthServiceImpl authServiceImpl;
 
     @PostMapping("/login")
     @Operation(summary = "Login User")
     public ResponseEntity<Response<LoginResponseDTO>> login(@RequestBody @Valid LoginDTO loginRequest) {
-        return GoodResponse.ok(authService.login(loginRequest));
+        return GoodResponse.ok(authServiceImpl.login(loginRequest));
     }
 
     @PostMapping("/register")
     @Operation(summary = "Register User")
     public ResponseEntity<Response<UserDTO>> registerDTO(@RequestBody @Valid RegisterDTO dto) {
-        return GoodResponse.ok(authService.register(dto));
+        return GoodResponse.ok(authServiceImpl.register(dto));
     }
 
     @PostMapping("/refresh")
     @Operation(summary = "Refresh JWTToken")
     public ResponseEntity<Response<String>> refresh(HttpServletRequest request) {
-        return GoodResponse.ok(authService.refresh(request));
+        return GoodResponse.ok(authServiceImpl.refresh(request));
     }
 }
